requirements.txt
streamlit
spacy
nltk
pandas
numpy
scikit-learn
openai
python-dotenv
openpyxl
sentence-transformers 
setuptools